Initial revision of meta-hosted-mender-login layer
This layer provides a graphical login to Hosted Mender, written in
PyQt5 and thus depends on meta-qt5, that retrieves and stores the
Hosted Mender tenant token used to interface with Mender API. The
application is run each time the unit boots. Source code is currently
self contained in the layer.

I've also included some recipes for configuring ethernet.

Signed-off-by: Alf-Rune <alf.rune@northern.tech>